requirements.txt
streamlit 
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-community
sentence-transformers
pdfplumber
langchain-huggingface